+++ gsoap/import/xlink.h
@@ -50,3 +50,4 @@ A commercial use license is available from Genivia, Inc., contact@genivia.com
 */
 
 //gsoap xlink schema import: http://www.w3.org/1999/xlink
+//gsoap xlink schema form:   unqualified